demo_web/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
sseclient-py>=1.7.2
pandas>=2.0.0 